# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ДомДело API
  version: 0.1.0
  description: API MVP для коллективных дел жителей дома
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session
    demoUser:
      type: apiKey
      in: header
      name: X-Demo-User
...
paths:
...
  /api/reports/house:
    get:
      tags:
        - reports
      security:
        - bearerAuth: []
        - demoUser: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - houseId
                  - address
                  - asOf
                  - totalCases
                  - openCases
                  - overdueForecasts
                  - manyConfirmed
                  - categories
                  - source
                properties:
                  houseId:
                    format: uuid
                    type: string
                  address:
                    type: string
                  asOf:
                    format: date-time
                    type: string
                  totalCases:
                    minimum: 0
                    type: integer
                  openCases:
                    minimum: 0
                    type: integer
                  overdueForecasts:
                    minimum: 0
                    type: integer
                  manyConfirmed:
                    minimum: 0
                    type: integer
                  categories:
                    type: array
                    items:
                      type: object
                      required:
                        - category
                        - cases
                        - confirmations
                      properties:
                        category:
                          anyOf:
                            - type: string
                              enum:
                                - lighting
                            - type: string
                              enum:
                                - entrance
                            - type: string
                              enum:
                                - elevator
                            - type: string
                              enum:
                                - water
                            - type: string
                              enum:
                                - heating
                            - type: string
                              enum:
                                - yard
                            - type: string
                              enum:
                                - other
                        cases:
                          minimum: 0
                          type: integer
                        confirmations:
                          minimum: 0
                          type: integer
                  source:
                    type: string
                    enum:
                      - domdelo_internal
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                  - message
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  requestId:
                    type: string
...
  /api/cases/{caseId}:
    get:
      tags:
        - cases
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: caseId
          required: true
      security:
        - bearerAuth: []
        - demoUser: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - number
                  - houseId
                  - title
                  - description
                  - category
                  - place
                  - status
                  - confirmationsCount
                  - watchersCount
                  - isWatched
                  - responsibleOrganization
                  - version
                  - isDemo
                  - createdAt
                  - updatedAt
                  - history
                  - attachments
                properties:
                  id:
                    format: uuid
                    type: string
                  number:
                    minimum: 1
                    type: integer
                  houseId:
                    format: uuid
                    type: string
                  title:
                    type: string
                  description:
                    type: string
                  category:
                    anyOf:
                      - type: string
                        enum:
                          - lighting
                      - type: string
                        enum:
                          - entrance
                      - type: string
                        enum:
                          - elevator
                      - type: string
                        enum:
                          - water
                      - type: string
                        enum:
                          - heating
                      - type: string
                        enum:
                          - yard
                      - type: string
                        enum:
                          - other
                  entrance:
                    type: string
                  place:
                    type: string
                  status:
                    anyOf:
                      - type: string
                        enum:
                          - draft
                      - type: string
                        enum:
                          - registered
                      - type: string
                        enum:
                          - assigned
                      - type: string
                        enum:
                          - in_progress
                      - type: string
                        enum:
                          - awaiting_resident_verification
                      - type: string
                        enum:
                          - resolved
                      - type: string
                        enum:
                          - disputed
                  confirmationsCount:
                    minimum: 1
                    type: integer
                  watchersCount:
                    minimum: 0
                    type: integer
                  isWatched:
                    type: boolean
                  responsibleOrganization:
                    type: string
                  assignee:
                    type: string
                  resultComment:
                    type: string
                  plannedCompletionAt:
                    format: date-time
                    type: string
                  version:
                    minimum: 1
                    type: integer
                  isDemo:
                    type: boolean
                  createdAt:
                    format: date-time
                    type: string
                  updatedAt:
                    format: date-time
                    type: string
                  history:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - toStatus
                        - actorName
                        - createdAt
                      properties:
                        id:
                          format: uuid
                          type: string
                        fromStatus:
                          anyOf:
                            - type: string
                              enum:
                                - draft
                            - type: string
                              enum:
                                - registered
                            - type: string
                              enum:
                                - assigned
                            - type: string
                              enum:
                                - in_progress
                            - type: string
                              enum:
                                - awaiting_resident_verification
                            - type: string
                              enum:
                                - resolved
                            - type: string
                              enum:
                                - disputed
                        toStatus:
                          anyOf:
                            - type: string
                              enum:
                                - draft
                            - type: string
                              enum:
                                - registered
                            - type: string
                              enum:
                                - assigned
                            - type: string
                              enum:
                                - in_progress
                            - type: string
                              enum:
                                - awaiting_resident_verification
                            - type: string
                              enum:
                                - resolved
                            - type: string
                              enum:
                                - disputed
                        comment:
                          type: string
                        actorName:
                          type: string
                        plannedCompletionAt:
                          format: date-time
                          type: string
                        createdAt:
                          format: date-time
                          type: string
                  attachments:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - kind
                        - url
                        - fileName
                        - mimeType
                        - createdAt
                      properties:
                        id:
                          format: uuid
                          type: string
                        kind:
                          anyOf:
                            - type: string
                              enum:
                                - problem
                            - type: string
                              enum:
                                - result
                        url:
                          type: string
                        fileName:
                          type: string
                        mimeType:
                          type: string
                        createdAt:
                          format: date-time
                          type: string
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                  - message
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  requestId:
                    type: string
  /api/cases/deduplication:
    post:
      tags:
        - cases
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - description
                - category
                - place
              properties:
                description:
                  minLength: 5
                  maxLength: 2000
                  type: string
                category:
                  anyOf:
                    - type: string
                      enum:
                        - lighting
                    - type: string
                      enum:
                        - entrance
                    - type: string
                      enum:
                        - elevator
                    - type: string
                      enum:
                        - water
                    - type: string
                      enum:
                        - heating
                    - type: string
                      enum:
                        - yard
                    - type: string
                      enum:
                        - other
                entrance:
                  maxLength: 20
                  type: string
                place:
                  minLength: 2
                  maxLength: 120
                  type: string
      security:
        - bearerAuth: []
        - demoUser: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - id
                    - number
                    - houseId
                    - title
                    - description
                    - category
                    - place
                    - status
                    - confirmationsCount
                    - watchersCount
                    - isWatched
                    - responsibleOrganization
                    - version
                    - isDemo
                    - createdAt
                    - updatedAt
                    - history
                    - attachments
                  properties:
                    id:
                      format: uuid
                      type: string
                    number:
                      minimum: 1
                      type: integer
                    houseId:
                      format: uuid
                      type: string
                    title:
                      type: string
                    description:
                      type: string
                    category:
                      anyOf:
                        - type: string
                          enum:
                            - lighting
                        - type: string
                          enum:
                            - entrance
                        - type: string
                          enum:
                            - elevator
                        - type: string
                          enum:
                            - water
                        - type: string
                          enum:
                            - heating
                        - type: string
                          enum:
                            - yard
                        - type: string
                          enum:
                            - other
                    entrance:
                      type: string
                    place:
                      type: string
                    status:
                      anyOf:
                        - type: string
                          enum:
                            - draft
                        - type: string
                          enum:
                            - registered
                        - type: string
                          enum:
                            - assigned
                        - type: string
                          enum:
                            - in_progress
                        - type: string
                          enum:
                            - awaiting_resident_verification
                        - type: string
                          enum:
                            - resolved
                        - type: string
                          enum:
                            - disputed
                    confirmationsCount:
                      minimum: 1
                      type: integer
                    watchersCount:
                      minimum: 0
                      type: integer
                    isWatched:
                      type: boolean
                    responsibleOrganization:
                      type: string
                    assignee:
                      type: string
                    resultComment:
                      type: string
                    plannedCompletionAt:
                      format: date-time
                      type: string
                    version:
                      minimum: 1
                      type: integer
                    isDemo:
                      type: boolean
                    createdAt:
                      format: date-time
                      type: string
                    updatedAt:
                      format: date-time
                      type: string
                    history:
                      type: array
                      items:
                        type: object
                        required:
                          - id
                          - toStatus
                          - actorName
                          - createdAt
                        properties:
                          id:
                            format: uuid
                            type: string
                          fromStatus:
                            anyOf:
                              - type: string
                                enum:
                                  - draft
                              - type: string
                                enum:
                                  - registered
                              - type: string
                                enum:
                                  - assigned
                              - type: string
                                enum:
                                  - in_progress
                              - type: string
                                enum:
                                  - awaiting_resident_verification
                              - type: string
                                enum:
                                  - resolved
                              - type: string
                                enum:
                                  - disputed
                          toStatus:
                            anyOf:
                              - type: string
                                enum:
                                  - draft
                              - type: string
                                enum:
                                  - registered
                              - type: string
                                enum:
                                  - assigned
                              - type: string
                                enum:
                                  - in_progress
                              - type: string
                                enum:
                                  - awaiting_resident_verification
                              - type: string
                                enum:
                                  - resolved
                              - type: string
                                enum:
                                  - disputed
                          comment:
                            type: string
                          actorName:
                            type: string
                          plannedCompletionAt:
                            format: date-time
                            type: string
                          createdAt:
                            format: date-time
                            type: string
                    attachments:
                      type: array
                      items:
                        type: object
                        required:
                          - id
                          - kind
                          - url
                          - fileName
                          - mimeType
                          - createdAt
                        properties:
                          id:
                            format: uuid
                            type: string
                          kind:
                            anyOf:
                              - type: string
                                enum:
                                  - problem
                              - type: string
                                enum:
                                  - result
                          url:
                            type: string
                          fileName:
                            type: string
                          mimeType:
                            type: string
                          createdAt:
                            format: date-time
                            type: string
...
servers:
  - url: http://localhost:8080
...